Spinning out `entity::Bullet` module and moving init code there
