+++ ports/amethyst/src/defender/entity/mod.rs
@@ -1,46 +1,13 @@
 use amethyst::{
-    ecs::prelude::{ Component, DenseVecStorage, Entity },
-    renderer::{
-        Material,
-        MeshHandle,
-    }
+    ecs::prelude::Entity,
 };
 
+mod bullet;
+pub use bullet::*;
 mod enemy;
 pub use enemy::*;
 mod player;
 pub use player::*;
-
-#[derive(Debug, Default)]
-pub struct Bullet {
-    /// Direction the bullet is going.
-    pub direction: f32,
-    /// Time-to-live for the bullet.
-    pub ttl: f32,
-}
-
-impl Bullet {
-    pub fn new(shooter: &Player) -> Self {
-        Bullet {
-            direction: shooter.direction,
-            ttl: 0.0
-        }
-    }
-}
-
-impl Component for Bullet {
-    type Storage = DenseVecStorage<Self>;
-}
-
-#[derive(Clone)]
-pub struct BulletResource {
-    pub material: Material,
-    pub mesh: MeshHandle
-}
-
-impl Component for BulletResource {
-    type Storage = DenseVecStorage<Self>;
-}
 
 #[derive(Debug, Default)]
 pub struct ScoreBoard {
